+++ examples/sv/elements/virtual_cls.sv
@@ -35,22 +35,6 @@ virtual class Packet #(WIDTH = 2) extends hello implements aa;
 endclass
 
 
-// interface class PutImp#(type PUT_T = logic); 
-
-//     pure virtual function void put(PUT_T a);
-//     endfunction
-
-// endclass
-
-// interface class Messaging #(type T = logic);
-//   pure virtual task          put(T t);
-//   pure virtual task          get(output T t);
-//   pure virtual task          peek(output T t);
-//   pure virtual function bit  try_peek(output T t);
-//   pure virtual function bit  try_put(T t);
-//   pure virtual function bit  try_get(output T t);
-// endclass
-
 module example_mod(
     `ifdef MY_VAR
         input wire clk,