+++ src/core/lang/sv/symbols/package.rs
@@ -22,7 +22,7 @@ use crate::core::lang::{
     reference::RefSet,
     sv::{
         error::SystemVerilogError,
-        token::{identifier::Identifier, keyword::Keyword, token::SystemVerilogToken},
+        token::{identifier::Identifier, keyword::Keyword, operator::Operator, token::SystemVerilogToken},
     },
     verilog::symbols::VerilogSymbol,
 };
@@ -78,7 +78,10 @@ impl Package {
         };
 
         // take terminator ';'
-        tokens.next().take().unwrap();
+        let t = tokens.next().take().unwrap();
+        if t.as_type().check_delimiter(&Operator::Terminator) == false {
+            return Err(SystemVerilogError::ExpectingOperator(Operator::Terminator))
+        }
 
         let mut refs = RefSet::new();